pyproject.toml
[tool.poetry]
name = "whisper-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.12.0"
pyaudio = "^0.2.14"
webrtcvad = "^2.0.10"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"
cython = "^3.0.8"
reazonspeech-espnet-asr = {path = "Reazonspeech/pkg/espnet-asr"}
typeguard = "2.13.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.mypy]
show_error_context = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true
warn_redundant_casts = true